Workaround changes in neovim
Neovim changed the way processes launched by system() are handled such
that calls to `tty` result in `/dev/tty: No such device or address`.
This broke oscyank.vim.

This fix uses the hack in
https://github.com/neovim/neovim/issues/8450#issuecomment-407063983 to
use /proc/$$PID to get a valid tty when `tty` fails.

With this change oscyank.vim works when the remote host is /proc/-based
(i.e. Linux, but not macOS, for example).